Develop SaaS -> Master SaaS

+++ Servers/utils/controlCategory.utils.ts
@@ -1,17 +1,17 @@
-import {
-  ControlCategory,
-  ControlCategoryModel,
-} from "../domain.layer/models/controlCategory/controlCategory.model";
+import { ControlCategoryModel } from "../domain.layer/models/controlCategory/controlCategory.model";
 import { sequelize } from "../database/db";
 import { createNewControlsQuery } from "./control.utils";
 import { ControlCategories } from "../structures/EU-AI-Act/compliance-tracker/controlCategories.struct";
 import { QueryTypes, Transaction } from "sequelize";
+import { IControlCategory } from "../domain.layer/interfaces/i.controlCategory";
 
-export const getAllControlCategoriesQuery = async (): Promise<
-  ControlCategory[]
+export const getAllControlCategoriesQuery = async (
+  tenant: string
+): Promise<
+  ControlCategoryModel[]
 > => {
   const controlCategories = await sequelize.query(
-    "SELECT * FROM controlcategories ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".controlcategories ORDER BY created_at DESC, id ASC`,
     {
       mapToModel: true,
       model: ControlCategoryModel,
@@ -21,10 +21,11 @@ export const getAllControlCategoriesQuery = async (): Promise<
 };
 
 export const getControlCategoryByIdQuery = async (
-  id: number
-): Promise<ControlCategory | null> => {
+  id: number,
+  tenant: string
+): Promise<IControlCategory | null> => {
   const result = await sequelize.query(
-    "SELECT * FROM controlcategories WHERE id = :id",
+    `SELECT * FROM "${tenant}".controlcategories WHERE id = :id`,
     {
       replacements: { id },
       mapToModel: true,
@@ -37,7 +38,7 @@ export const getControlCategoryByIdQuery = async (
 export const getControlCategoryByTitleAndProjectIdQuery = async (
   title: string,
   projectId: number
-): Promise<ControlCategory | null> => {
+): Promise<IControlCategory | null> => {
   const result = await sequelize.query(
     "SELECT * FROM controlcategories WHERE title = :title AND project_id = :project_id",
     {
@@ -50,10 +51,11 @@ export const getControlCategoryByTitleAndProjectIdQuery = async (
 };
 
 export const getControlCategoryByProjectIdQuery = async (
-  projectId: number
-): Promise<ControlCategory[]> => {
+  projectId: number,
+  tenant: string
+): Promise<IControlCategory[]> => {
   const result = await sequelize.query(
-    "SELECT * FROM controlcategories WHERE project_id = :project_id ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".controlcategories WHERE project_id = :project_id ORDER BY created_at DESC, id ASC`,
     {
       replacements: { project_id: projectId },
       mapToModel: true,
@@ -64,11 +66,12 @@ export const getControlCategoryByProjectIdQuery = async (
 };
 
 export const createControlCategoryQuery = async (
-  controlCategory: ControlCategory,
+  controlCategory: ControlCategoryModel,
+  tenant: string,
   transaction: Transaction
-): Promise<ControlCategory> => {
+): Promise<ControlCategoryModel> => {
   const result = await sequelize.query(
-    `INSERT INTO controlcategories (
+    `INSERT INTO "${tenant}".controlcategories (
       project_id, title, order_no
     ) VALUES (:project_id, :title, :order_no) RETURNING *`,
     {
@@ -88,25 +91,28 @@ export const createControlCategoryQuery = async (
 
 export const updateControlCategoryByIdQuery = async (
   id: number,
-  controlCategory: Partial<ControlCategory>,
+  controlCategory: Partial<ControlCategoryModel>,
+  tenant: string,
   transaction: Transaction
-): Promise<ControlCategory | null> => {
-  const updateControlCategory: Partial<Record<keyof ControlCategory, any>> = {};
+): Promise<ControlCategoryModel | null> => {
+  const updateControlCategory: Partial<
+    Record<keyof ControlCategoryModel, any>
+  > = {};
   const setClause = ["title"]
     .filter((f) => {
       if (
-        controlCategory[f as keyof ControlCategory] !== undefined &&
-        controlCategory[f as keyof ControlCategory]
+        controlCategory[f as keyof ControlCategoryModel] !== undefined &&
+        controlCategory[f as keyof ControlCategoryModel]
       ) {
-        updateControlCategory[f as keyof ControlCategory] =
-          controlCategory[f as keyof ControlCategory];
+        updateControlCategory[f as keyof ControlCategoryModel] =
+          controlCategory[f as keyof ControlCategoryModel];
         return true;
       }
     })
     .map((f) => `${f} = :${f}`)
     .join(", ");
 
-  const query = `UPDATE controlcategories SET ${setClause} WHERE id = :id RETURNING *;`;
+  const query = `UPDATE "${tenant}".controlcategories SET ${setClause} WHERE id = :id RETURNING *;`;
 
   updateControlCategory.id = id;
 
@@ -123,10 +129,11 @@ export const updateControlCategoryByIdQuery = async (
 
 export const deleteControlCategoryByIdQuery = async (
   id: number,
+  tenant: string,
   transaction: Transaction
 ): Promise<Boolean> => {
   const result = await sequelize.query(
-    "DELETE FROM controlcategories WHERE id = :id RETURNING *",
+    `DELETE FROM "${tenant}".controlcategories WHERE id = :id RETURNING *`,
     {
       replacements: { id },
       mapToModel: true,
@@ -141,10 +148,11 @@ export const deleteControlCategoryByIdQuery = async (
 export const createNewControlCategories = async (
   projectId: number,
   enable_ai_data_insertion: boolean,
+  tenant: string,
   transaction: Transaction
 ) => {
   const createdControlCategories = [];
-  let query = `INSERT INTO controlcategories(
+  let query = `INSERT INTO "${tenant}".controlcategories(
     project_id, title, order_no
   ) VALUES (:project_id, :title, :order_no) RETURNING *;`;
   for (let controlCategoryStruct of ControlCategories) {
@@ -164,6 +172,7 @@ export const createNewControlCategories = async (
       control_category_id,
       controlCategoryStruct.controls,
       enable_ai_data_insertion,
+      tenant,
       transaction
     );
     createdControlCategories.push({ ...result[0].dataValues, controls });
